Issue #214 -  NanoHTTPD uri router. Single file
